refactor: Split UserProfile into smaller components and optimize performance
