--- src/components/Panel/UserProfile/types.ts
@@ -1,8 +1,0 @@
-export type UserData = {
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  company: string;
-  nip: string;
-};
