+++ src/components/UserProfile/hooks/useUserData.hook.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { API_CONFIG } from '../../../../config/api.config';
+import { API_CONFIG } from '../../../config/api.config';
 
 export type UserData = {
   firstName: string;
